Adding a README.md file for the repository
When browsing projects on github.com any README file is shown prominently
o the user, so this is an important way to introduce the project.

This text is taken directly from the current introduction in graphs.py
with minor reformatting for GitHub's markup style, and minor rewording
since it is no longer part of the Python script.