For Python 3, no longer need to use 2to3 print fixer, or fix doctests
The remaining changes by 2to3 are trivial (e.g. wrapping iterators like
the range function with a list call), and could be handled in a single
codebase for both Python 2 and 3, with one exception: As long as we
continue to support Python 2.4 and 2.5 we cannot use the Python 3 style
exception catching available in Python 2.6 onwards.